Replace ad-hoc incorrect JSON parser with proper ujson parser
