+++ akka-http/client/src/main/scala/endpoints/akkahttp/client/BuiltInErrors.scala
@@ -1,12 +1,11 @@
 package endpoints.akkahttp.client
 
-import endpoints.algebra.InvalidCodec.invalidCodec
 import endpoints.{Invalid, algebra}
 
 trait BuiltInErrors extends algebra.BuiltInErrors { this: EndpointsWithCustomErrors =>
 
   def clientErrorsResponseEntity: ResponseEntity[Invalid] =
-    stringCodecResponse(invalidCodec)
+    stringCodecResponse(endpoints.ujson.codecs.invalidCodec)
 
   def serverErrorResponseEntity: ResponseEntity[Throwable] =
     mapResponseEntity(clientErrorsResponseEntity)(invalid => new Throwable(invalid.errors.mkString(". ")))
