+++ play/server/src/main/scala/endpoints/play/server/BuiltInErrors.scala
@@ -1,6 +1,5 @@
 package endpoints.play.server
 
-import endpoints.algebra.InvalidCodec.invalidCodec
 import endpoints.{Invalid, algebra}
 import play.api.http.{ContentTypes, Writeable}
 
@@ -8,7 +7,7 @@ trait BuiltInErrors extends algebra.BuiltInErrors { this: EndpointsWithCustomErr
 
   def clientErrorsResponseEntity: ResponseEntity[Invalid] = {
     val playCodec = implicitly[play.api.mvc.Codec]
-    Writeable(invalid => playCodec.encode(invalidCodec.encode(invalid)), Some(ContentTypes.JSON))
+    Writeable(invalid => playCodec.encode(endpoints.ujson.codecs.invalidCodec.encode(invalid)), Some(ContentTypes.JSON))
   }
 
   def serverErrorResponseEntity: ResponseEntity[Throwable] =
