+++ scalaj/build.sbt
@@ -2,6 +2,7 @@ import EndpointsSettings._
 
 val `algebra-jvm` = LocalProject("algebraJVM")
 val `algebra-circe-jvm` = LocalProject("algebra-circeJVM")
+val `openapi-jvm` = LocalProject("openapiJVM")
 
 val `scalaj-client` =
   project.in(file("client"))
@@ -13,5 +14,6 @@ val `scalaj-client` =
         "org.scalaj" %% "scalaj-http" % "2.4.2"
       )
     )
+    .dependsOn(`openapi-jvm`)
     .dependsOn(`algebra-jvm` % "test->test;compile->compile")
     .dependsOn(`algebra-circe-jvm` % "test->test")
