+++ sttp/build.sbt
@@ -7,7 +7,7 @@ val `sttp-client` =
   project.in(file("client"))
     .settings(
       publishSettings,
-      `scala 2.11 to latest`,
+      `scala 2.12 to latest`,
       name := "endpoints-sttp-client",
       libraryDependencies ++= Seq(
         "com.softwaremill.sttp" %% "core" % sttpVersion,
@@ -15,4 +15,5 @@ val `sttp-client` =
         "com.typesafe.akka" %% "akka-stream" % "2.5.26" % Test
       )
     )
+    .dependsOn(LocalProject("openapiJVM"))
     .dependsOn(`algebra-jvm` % "compile->compile;test->test", `algebra-playjson-jvm` % "test->test")
