+++ sttp/client/src/main/scala/endpoints/sttp/client/BuiltInErrors.scala
@@ -1,11 +1,11 @@
 package endpoints.sttp.client
 
-import endpoints.algebra.InvalidCodec.invalidCodec
 import endpoints.{Invalid, algebra}
 
 trait BuiltInErrors[R[_]] extends algebra.BuiltInErrors { this: EndpointsWithCustomErrors[R] =>
 
-  def clientErrorsResponseEntity: ResponseEntity[Invalid] = stringCodecResponse(invalidCodec)
+  def clientErrorsResponseEntity: ResponseEntity[Invalid] =
+    stringCodecResponse(endpoints.ujson.codecs.invalidCodec)
 
   def serverErrorResponseEntity: ResponseEntity[Throwable] =
     mapResponseEntity(clientErrorsResponseEntity)(invalid => new Throwable(invalid.errors.mkString(". ")))
